feat: add rich presence support and refactor event handling
- Added `rich_presence` feature with necessary dependencies in `Cargo.toml`
- Introduced `ChannelPlugin` and `ChannelRes` for shared event handling
- Refactored event handling to use `EventCollection` for both `bot` and `rich_presence` features
- Updated `DiscordBotPlugin` to use `ChannelRes` for event communication
- Added `DiscordRichPresencePlugin` for integrating Discord Rich Presence
- Created `EventHandler` for handling rich presence events
- Updated system setup to conditionally run based on `ChannelRes` existence
