+++ src/rich_presence/discord_sdk.rs
@@ -1,0 +1,2 @@
+#[doc(hidden)]
+pub use discord_sdk::*;
